Refactor VCL OpenGL support
* Adds basic SPIR-V support; include specialization constants
* Move GL context create into a custom context class; initialize using EGL
* Move resource allocation to GraphicsEngine; remove DynamicTexture
* Use GraphicsEngine for the grid-cube example, get rid of free uniforms, start using precompiled shaders
* Extend GraphicsEngine
** Buffer clear support
** Draw call support
** Type abstraction for GLSL UniformBuffer types
** Additional rasterization pipeline state

+++ src/tests/vcl.graphics.opengl/radixsort.cpp
@@ -49,7 +49,7 @@ void ExecuteRadixSortTest(unsigned int size)
 
 	Runtime::BufferDescription desc =
 	{
-		num_keys * sizeof(int),
+		num_keys * static_cast<unsigned int>(sizeof(int)),
 		Runtime::ResourceUsage::Staging,
 		Runtime::ResourceAccess::Read | Runtime::ResourceAccess::Write
 	};
@@ -57,7 +57,7 @@ void ExecuteRadixSortTest(unsigned int size)
 	Runtime::BufferInitData data =
 	{
 		numbers.data(),
-		num_keys * sizeof(int)
+		num_keys * static_cast<unsigned int>(sizeof(int))
 	};
 
 	auto keys = Vcl::make_owner<Runtime::OpenGL::Buffer>(desc, true, true, &data);
